Implementei exibir detalhes do usuario cadastrado, deletar, voltar, comecei a funcao editar, mas nao deu tenpo de terminar novo link do surge http://well-to-do-rest.surge.sh/

+++ semana7/labeusers/labe-users/src/components/Usuarios/Usuarios.js
@@ -2,6 +2,7 @@ import React from 'react'
 import axios from 'axios'
 import Styled from 'styled-components'
 import UsuarioImg from '../../img/usuarios.jpg'
+import DetalheUsuario from '../Usuarios/DetalheUsuario'
 
 
 const UsuarioDiv = Styled.div`
@@ -35,11 +36,12 @@ const UsuarioHeadCol1 = Styled.div`
 	border: 1px solid #888;
 	text-transform:uppercase;
 	color: #fff;
-	width:68%;
+	width:calc(68% - 20px);
 	height:100%;
 	display:flex;
 	align-items:center;
-	justify-content:center;
+	padding: 0 10px;
+	justify-content:space-between;
 `;
 const UsuarioHeadCol2 = Styled.div`
 	background:#888;
@@ -67,11 +69,12 @@ const UsuarioTh1 = Styled.div`
 	border: 1px solid #888;
 	text-transform:uppercase;
 	color: #888;
-	width:68%;
+	width:calc(68% - 20px);
 	height:100%;
 	display:flex;
 	align-items:center;
-	justify-content:center;
+	justify-content:space-between;
+	padding: 0 10px;
 	background:#fff;
 `;
 const UsuarioTh2 = Styled.div`
@@ -117,11 +120,27 @@ const UsuariosCenter = Styled.div`
 	align-items:center;
 	padding-bottom:30px;
 `;
+const UsuarioSpan = Styled.span`
+	color: #2aaa7e;
+	font-weight:700;
+	font-size:20px;
+	cursor:pointer;
+	text-align:center;
+	&:hover{
+		color:#17efa4;
+	}
+`;
 
 
 class Usuarios extends React.Component{
 	state = {
-		listaDeUsuarios:[]
+		listaDeUsuarios:[],
+		detalhesUsuario:'',
+		showDetalhesUsuario:false,
+		editarNome:true,
+		botaoEditarNome:true,
+		editarEmail:true,
+		botaoEditarEmail:true
 	}
 
 	componentDidMount = () => {
@@ -143,19 +162,21 @@ class Usuarios extends React.Component{
 	}
 
 	deletarUsuario = (iD) => {
-		console.log("id pelo click",iD)
-		axios.delete(`https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users/${iD}`,{
+		const confirmaExclusao = window.confirm("Tem certeza que deseja deletar esse usuário? ")
+		if(confirmaExclusao){
+			axios.delete(`https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users/${iD}`,{
 			headers: {
 				Authorization: "cristiane-da-rocha-dumont",
 				id:iD
 			}
-		}).then((resposta) => {
-			alert('Usuario: ' + resposta.data.users.name + "deletado com sucesso !")
-			console.log(resposta)
-		}).catch((error) => {
-			console.log(error)
-		})
-
+			}).then((resposta) => {
+				alert("Usuario deletado com sucesso!")
+				console.log(resposta)
+			}).catch((error) => {
+				console.log(error)
+			})
+		}
+		this.setState({showDetalhesUsuario:false})
 	}
 	componentDidUpdate = () => {
 		axios.get("https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users",{
@@ -170,11 +191,100 @@ class Usuarios extends React.Component{
 			console.log("Este Erro aqui",error)
 		})
 	}
+
+	openDetalhes = (iD) => {
+		axios.get(`https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users/${iD}`,{
+			headers: {
+				Authorization: "cristiane-da-rocha-dumont",
+				id:iD
+			}
+		}).then((resposta) => {
+			this.setState({
+				detalhesUsuario: {
+					name: resposta.data.name,
+					id: resposta.data.id,
+					email: resposta.data.email
+				}
+			})
+		}).catch((error) => {
+			console.log(error)
+		})
+		this.setState({showDetalhesUsuario:true})
+	}
+	fecharDetalhes = () => {
+		this.setState({showDetalhesUsuario:false})
+	}
+	abrirEditarNome = () => {
+		this.setState({
+			editarNome:false,
+			botaoEditarNome:false
+		})
+	}
+	abrirEditarEmail = () => {
+		this.setState({
+			editarEmail:false,
+			botaoEditarEmail:false
+		})
+		console.log('chegou aqui')
+	}
+	mudarNome = (event) => {
+		if(this.state.editarNome === false){
+			this.setState({detalhesUsuario:{
+			nome: event.target.value
+		}})
+		}		
+	}
+	mudarEmail = (event) => {
+		if(this.state.editarEmail === false){
+			this.setState({detalhesUsuario:{
+			email: event.target.value
+		}})
+		}		
+	}
+	salvarNome = (iD) => {
+		console.log("Cris")
+		const body = {
+			name: this.state.detalhesUsuario.nome,
+			email: this.state.detalhesUsuario.email
+		}
+		console.log(body,iD)
+		axios.put(`https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users//${iD}`,body,{
+		headers: {
+			Authorization: "cristiane-da-rocha-dumont",
+			id:iD
+		}
+		}).then((resposta) => {
+			alert("Usuario Salvo com sucesso!")
+			console.log(resposta)
+		}).catch((error) => {
+			alert("deu erro")
+			console.log(error)
+		})
+
+	}
 	render(){
+		let usuarioDetalhes = ''
+		if(this.state.showDetalhesUsuario === true){
+			usuarioDetalhes = <DetalheUsuario
+								nomeDoUsuario={this.state.detalhesUsuario.name}
+								emailDoUsuario={this.state.detalhesUsuario.email}
+								botaoDeletar={() => this.deletarUsuario(this.state.detalhesUsuario.id) }
+								fecharDetalhes={this.fecharDetalhes}
+								editarNome={this.state.editarNome}
+								abrirEditarNome={this.abrirEditarNome}
+								abrirEditarEmail={this.abrirEditarEmail}
+								mudarNome={this.mudarNome}
+								mudarEmail={this.mudarEmail}
+								mudarBotaoNome={this.state.botaoEditarNome}
+								mudarBotaoEmail={this.state.botaoEditarEmail}
+								salvarNome={() => this.salvarNome(this.state.detalhesUsuario.id) }
+							  />
+		}
+
 		let usuarios = this.state.listaDeUsuarios.map((usuario) => {
 			return(
 				<UsuarioTr key={usuario.id}>
-						<UsuarioTh1>{usuario.name}</UsuarioTh1>
+						<UsuarioTh1>{usuario.name}<UsuarioSpan onClick={() => this.openDetalhes(usuario.id)}> + </UsuarioSpan></UsuarioTh1>
 						<UsuarioTh2><UsuarioDeleteButton onClick={() => this.deletarUsuario(usuario.id)}> x </UsuarioDeleteButton></UsuarioTh2>
 				</UsuarioTr>
 			)
@@ -183,9 +293,10 @@ class Usuarios extends React.Component{
 			<UsuarioDiv>
 				<UsuariosCenter>
 				    <UsuarioH2>Usuários Cadastrados </UsuarioH2>
+				    {usuarioDetalhes}
 					<UsuarioTabela>
 							<UsuarioHead>
-								<UsuarioHeadCol1>Nome Cadastrado</UsuarioHeadCol1>
+								<UsuarioHeadCol1>Nome Cadastrado<span>Detalhes</span></UsuarioHeadCol1>
 								<UsuarioHeadCol2>Deletar Usuario</UsuarioHeadCol2>
 							</UsuarioHead>
 						<UsuarioTbody>
